requirements.txt
amqp==5.1.1
celery==5.2.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-admin-interface==0.22.1
django-colorfield==0.7.2
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
paho-mqtt==1.6.1
Pillow==9.3.0
psycopg2-binary==2.9.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
requests==2.27.1
redis==4.0.2
sendgrid==6.9.7
six==1.16.0
eventlet==0.29.0
numpy
mongoengine==0.20.0
pylint-mongoengine==0.4.0
pymongo==3.11.0
djongo
